% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_query.R
\name{get_query}
\alias{get_query}
\title{Parse JSON resulting from AQL query to data frame}
\usage{
get_query(baseURL, credentials, aql_query, full_path = FALSE)
}
\arguments{
\item{baseURL}{url addres of the REST service (character string).}

\item{credentials}{character vector of the authentication pair of \code{username} and \code{password}.}

\item{aql_query}{character string containing the AQL query.}

\item{full_path}{logical value indicating the degree of flattening. If TRUE each value (even repetitions on the same level) in the JSON is assigned a unique identifier (basically flattened address). If FALSE values are sorted in the outpt data frame according to JSON path, case number and unique sequence within the same case (if there are several repetitions on the same JSON level).}
}
\value{
A data frame containing the parsed JSON result.
}
\description{
This function applies the GET method on the selected AQL query to obtain data from EhrScape platform.
}
\examples{
baseURL <- "https://rest.ehrscape.com/rest/v1/"
aql_query <- "select 
   a_a/data[at0002|History|]/events[at0003|Any event|]/data[at0001]/items[at0004|Temperature|]/value 
   as Temperature, 
   a_b/data[at0001|history|]/events[at0006|any event|]/data[at0003]/items[at0004|Systolic|]/value 
   as Systolic, 
   a_b/data[at0001|history|]/events[at0006|any event|]/data[at0003]/items[at0005|Diastolic|]/value 
   as Diastolic 
   from EHR e contains COMPOSITION 
   a contains (OBSERVATION a_a[openEHR-EHR-OBSERVATION.body_temperature.v1] 
   and OBSERVATION a_b[openEHR-EHR-OBSERVATION.blood_pressure.v1]) offset 0 limit 100"
query_data <- get_query(baseURL, credentials = c("guidemo", "gui?!demo123"), aql_query, full_path = FALSE)
}

