% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query2csv.R
\name{get_query_csv}
\alias{get_query_csv}
\title{Apply the GET query/csv service from EHRScape to import query results as table}
\usage{
get_query_csv(baseURL, credentials = c(user_name, password), aql_query)
}
\arguments{
\item{baseURL}{url addres of the REST service (character string).}

\item{credentials}{character vector of the authentication pair of \code{username} and \code{password}.}

\item{aql_query}{character string containing the AQL query.}
}
\value{
A data frame containing the JSON content of query results.
}
\description{
This function applies the GET method on the selected AQL query to obtain data in CSV format from EhrScape platform.
}
\examples{
baseURL <- "https://rest.ehrscape.com/rest/v1/"
aql_query <- "select 
   a_a/data[at0002|History|]/events[at0003|Any event|]/data[at0001]/items[at0004|Temperature|]/value 
   as Temperature, 
   a_b/data[at0001|history|]/events[at0006|any event|]/data[at0003]/items[at0004|Systolic|]/value 
   as Systolic, 
   a_b/data[at0001|history|]/events[at0006|any event|]/data[at0003]/items[at0005|Diastolic|]/value 
   as Diastolic from EHR e contains COMPOSITION 
   a contains (OBSERVATION a_a[openEHR-EHR-OBSERVATION.body_temperature.v1] 
   and OBSERVATION a_b[openEHR-EHR-OBSERVATION.blood_pressure.v1]) offset 0 limit 100"
query_data <- get_query_csv(baseURL, credentials = c("guidemo", "gui?!demo123"), aql_query)
}

