% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_composition.R
\name{post_composition}
\alias{post_composition}
\title{POST data to EHRScape composition}
\usage{
post_composition(baseURL, credentials, templateId, ehrId, format, composition)
}
\arguments{
\item{baseURL}{url addres of the REST service (character string).}

\item{credentials}{character vector of the authentication pair of \code{username} and \code{password}.}

\item{templateId}{character string containing the identifier of one of the existing templates in the EhrScape platform.}

\item{ehrId}{character string containing the identifier of an existing EHR record.}

\item{format}{character string indicating whether the provided JSON composition is in structured ("STRUCTURED") or flattened ("FLAT") form.}

\item{composition}{the OpenEhr composition in JSON format to be saved to EhrScape platform. It can be given as character string or as a structured list.}
}
\value{
A list of the full server response and the parsed content field containing the link to the created composition.
}
\description{
This function applies the GET method on the selected AQL query to obtain data from EhrScape platform.
}
\examples{
baseURL <- "https://rest.ehrscape.com/rest/v1/"
templateId <- "Vital Signs"
ehrId <- "6e031066-14df-46db-8320-bce31199fcbd"
format <- "FLAT"
composition <- '{"ctx/language": "en", "ctx/territory": "US", 
   "ctx/composer_name": "Silvia Blake", 
   "ctx/time": "2015-12-03T14:55:51.868+01:00", 
   "ctx/id_namespace": "HOSPITAL-NS", "ctx/id_scheme": "HOSPITAL-NS", 
   "ctx/health_care_facility|name": "Hospital", 
   "ctx/health_care_facility|id": "9091", 
   "vital_signs/body_temperature:0/any_event:0/temperature|magnitude": 37.94, 
   "vital_signs/body_temperature:0/any_event:0/temperature|unit": "°C"}'
post_c <- post_composition(baseURL, credentials = c("guidemo", "gui?!demo123"), templateId, ehrId, format, composition)
}

